[src/MaxText/utils/ckpt_conversion] Split into one file per ckpt_conversion following strategy pattern ; [tests/ckpt_conversion_param_mapping_strategy_test.py] Add test
